Tranformando valor de dinheiro em decimal

+++ app/database/migrations/0000_00_00_000013_setup_offers_table.php
@@ -33,8 +33,8 @@ class SetupOffersTable extends Migration {
 		$table->string('title')->nullable();
 		$table->string('subtitle')->nullable();
 		$table->string('subsubtitle')->nullable(); // HTML
-		$table->integer('price_original')->nullable();
-		$table->integer('price_with_discount')->nullable();
+		$table->decimal('price_original', 7, 2)->nullable();
+		$table->decimal('price_with_discount', 7, 2)->nullable();
 		$table->text('rules')->nullable(); // HTML
 		$table->text('features')->nullable(); // HTML
 		$table->string('saveme_title')->nullable();
